+++ src/prefect/serialization/schedule.py
@@ -7,6 +7,7 @@ import prefect
 from prefect.serialization import schedule_compat
 from prefect.utilities.serialization import (
     DateTimeTZ,
+    JSONCompatible,
     ObjectSchema,
     OneOfSchema,
     StatefulFunctionReference,
@@ -38,6 +39,9 @@ class IntervalClockSchema(ObjectSchema):
     start_date = DateTimeTZ(allow_none=True)
     end_date = DateTimeTZ(allow_none=True)
     interval = fields.TimeDelta(precision="microseconds", required=True)
+    parameter_defaults = fields.Dict(
+        key=fields.Str(), values=JSONCompatible(), allow_none=True
+    )
 
     @post_dump
     def _interval_validation(self, data: dict, **kwargs: Any) -> dict:
@@ -67,6 +71,9 @@ class CronClockSchema(ObjectSchema):
     start_date = DateTimeTZ(allow_none=True)
     end_date = DateTimeTZ(allow_none=True)
     cron = fields.String(required=True)
+    parameter_defaults = fields.Dict(
+        key=fields.Str(), values=JSONCompatible(), allow_none=True
+    )
 
 
 class DatesClockSchema(ObjectSchema):
@@ -76,6 +83,9 @@ class DatesClockSchema(ObjectSchema):
     start_date = DateTimeTZ(allow_none=True)
     end_date = DateTimeTZ(allow_none=True)
     dates = DateTimeTZ(required=True, many=True)
+    parameter_defaults = fields.Dict(
+        key=fields.Str(), values=JSONCompatible(), allow_none=True
+    )
 
 
 class ClockSchema(OneOfSchema):